+++ src/iter/mod.rs
@@ -1,8 +1,11 @@
 //! Iterator parsing utilities.
 
 mod advance;
+mod next_until;
+mod next_while;
 mod peek;
 
 pub use advance::*;
-
+pub use next_until::*;
+pub use next_while::*;
 pub use peek::*;