+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat!: Introduce `NoMatchError` rather than `()` to be more clear.
 - feat: Implement `ParserIterator` wrapper
 - feat: Implement `advance_until` combinator
+- feat: Implement `next_while` and `next_until` combinators
 
 ## 0.2.0-alpha
 - feat: Implement `Group` parsers.